Add placeholder validation for login :hammer:
